+++ xarray/datatree_/datatree/tests/test_datatree.py
@@ -603,6 +603,13 @@ class TestAccess:
         var_keys = list(dt.variables.keys())
         assert all(var_key in key_completions for var_key in var_keys)
 
+    def test_operation_with_attrs_but_no_data(self):
+        # tests bug from xarray-datatree GH262
+        xs = xr.Dataset({"testvar": xr.DataArray(np.ones((2, 3)))})
+        dt = DataTree.from_dict({"node1": xs, "node2": xs})
+        dt.attrs["test_key"] = 1  # sel works fine without this line
+        dt.sel(dim_0=0)
+
 
 class TestRestructuring:
     def test_drop_nodes(self):